+++ CHANGES.md
@@ -2,6 +2,13 @@
 
 Versions are following semantic versioning (https://semver.org/) 
 
+
+## Version 0.0.2 (04.02.2022)
+
+- Added Linear Regression with SGD
+- Added Lasso, Ridge, ElasticNet regressions
+- Added L1, L2, L1 + L2 regularizations
+
 ## Version 0.0.1 (16.01.2022)
 
 - Added linear regression with OLS (Ordinary Least Square) solver